feat: update shared actions
